const_string_encryption: Don't create/modify static constructor when there's no code to add
Right now, even when there's no static string in a class, obfuscapk adds a static
constructor (which is empty because there's no decryption code to add) and every
single class in a project is modified because of this.

This patch skips adding the static constructor when there's no code to add
so only classes with static strings are modified.